requirements.txt
# Kroger API data engineering
# Install: pip install -r requirements.txt

# API ingestion
requests==2.32.3
python-dotenv==1.0.1

# FastAPI app
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
aiofiles==25.1.0

# Database
asyncpg==0.30.0        # FastAPI app (async)
psycopg2-binary==2.9.10  # ingestion scripts + backfill (sync)

# Data exploration (notebook)
pandas==2.2.3
ipykernel==6.29.5
